% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{calculate_responsibilities}
\alias{calculate_responsibilities}
\title{Calculates responsibilities for each point (or bin) at each time point, given parameter estimates for all clusters at all times}
\usage{
calculate_responsibilities(y, mu, Sigma, pi)
}
\arguments{
\item{y}{length T list with \code{y[[t]]} being a n_t-by-d matrix}

\item{mu}{a T-by-K-by-d array of means}

\item{Sigma}{a T-K-by-d-by-d array of covariance matrices}

\item{pi}{a T-by-K vector of probabilities}
}
\description{
Calculates responsibilities for each point (or bin) at each time point, given parameter estimates for all clusters at all times
}