% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{default_biomass}
\alias{default_biomass}
\title{Creates a biomass list of length T, where each element \code{biomass[[t]]} is a numeric vector of length n_t containing just 1's.}
\usage{
default_biomass(y)
}
\arguments{
\item{y}{length T list with \code{y[[t]]} being a n_t-by-d matrix}
}
\description{
Creates a biomass list of length T, where each element \code{biomass[[t]]} is a numeric vector of length n_t containing just 1's.
}