% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{flowkernel-package}
\alias{flowkernel}
\alias{flowkernel-package}
\title{Smoothly Varying Mixture of Gaussians Modeling}
\description{
This package uses kernel-smoothed EM to estimate a smoothly varying mixture of Gaussians model.
}
\author{
\strong{Maintainer}: Jacob Bien \email{jbien@usc.edu}

Authors:
\itemize{
  \item Farhad de Sousa \email{fdesousa@usc.edu}
}

}
\keyword{internal}