% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{init_bayes}
\alias{init_bayes}
\title{Initialize the Kernel EM-algorithm using Bayesian methods}
\usage{
init_bayes(
  y,
  K,
  biomass = default_biomass(y),
  num_iter = 1,
  lap_smooth_const = 0
)
}
\arguments{
\item{y}{length T list with \code{y[[t]]} being a n_t-by-d matrix}

\item{K}{number of components}

\item{biomass}{A list of length T, where each element \code{biomass[[t]]} is a numeric vector of length n_t containing the biomass (or count) of particles in each bin}

\item{num_iter}{number of iterations of EM to perform #need to think more about iterations for Bayes init}

\item{lap_smooth_const}{laplace smoothing constant #More explanation needed}
}
\description{
Initialize the Kernel EM-algorithm using Bayesian methods
}