% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{init_const}
\alias{init_const}
\title{Initialize the Kernel EM-algorithm using constant parameters}
\usage{
init_const(y, K, times_to_sample = 50, points_to_sample = 50)
}
\arguments{
\item{y}{length T list with \code{y[[t]]} being a n_t-by-d matrix}

\item{K}{number of components}

\item{times_to_sample}{number of time points to sample}

\item{points_to_sample}{number of bins to sample from each sampled time point}
}
\description{
Initialize the Kernel EM-algorithm using constant parameters
}