% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{kernel_em}
\alias{kernel_em}
\title{Kernel-smoothed EM algorithm}
\usage{
kernel_em(y, K, hmu, hSigma, hpi, num_iter = 100)
}
\arguments{
\item{y}{length T list with \code{y[[t]]} being a n_t-by-d matrix}

\item{K}{number of components}

\item{hmu}{bandwidth for mu parameter}

\item{hSigma}{bandwidth for Sigma parameter}

\item{hpi}{bandwidth for pi parameter}

\item{num_iter}{number of iterations of EM to perform}
}
\description{
Kernel-smoothed EM algorithm
}