% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_1d_means}
\alias{plot_1d_means}
\title{Plot cluster centers as 1-d projection over time}
\usage{
plot_1d_means(mu, dim = 1)
}
\arguments{
\item{mu}{a T-by-K-by-d array of means}

\item{dim}{specify which dimension to be plotted: 1 (diam), 2 (chl_small), or 3 (pe)}
}
\description{
Plot cluster centers as 1-d projection over time
}