% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_1d_means_triple}
\alias{plot_1d_means_triple}
\title{Plot cluster means as 1-d projection over time, with all three dimensions plotted together, in separate plots}
\usage{
plot_1d_means_triple(mu)
}
\arguments{
\item{mu}{a T-by-K-by-d array of means}
}
\description{
Plot cluster means as 1-d projection over time, with all three dimensions plotted together, in separate plots
}