% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_1d_means_with_width}
\alias{plot_1d_means_with_width}
\title{Plot cluster means as 1-d projection over time, with line widths determined by pi}
\usage{
plot_1d_means_with_width(mu, pi, dim = 1)
}
\arguments{
\item{mu}{a T-by-K-by-d array of means}

\item{pi}{A T-by-K array, with each row consisting of probabilities that sum to one}
}
\description{
Plot cluster means as 1-d projection over time, with line widths determined by pi
}