% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_3d_centers}
\alias{plot_3d_centers}
\title{Plot cluster centers in 3-d}
\usage{
plot_3d_centers(y, z, mu)
}
\arguments{
\item{y}{length T list with \code{y[[t]]} being a n_t-by-d matrix}

\item{z}{a length T list with \code{z[[t]]} being a n_t vector of cluster assignments}

\item{mu}{a T-by-K-by-d array of means}
}
\description{
Plot cluster centers in 3-d
}