% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot biomass over time for each cluster}
\usage{
plot_biomass(biomass, resp)
}
\arguments{
\item{biomass}{A list of length T, where each element \code{biomass[[t]]} is a numeric vector of length n_t containing the biomass (or count) of particles in each bin}

\item{resp}{length T list with \code{y[[t]]} being a n_t-by-K matrix}
}
\description{
Plot biomass over time for each cluster
}