% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_data_and_model}
\alias{plot_data_and_model}
\title{Plot data colored by cluster assignment with cluster means}
\usage{
plot_data_and_model(y, z, mu)
}
\arguments{
\item{y}{length T list with \code{y[[t]]} being a n_t-by-d matrix}

\item{z}{a length T list with \code{z[[t]]} being a n_t vector of cluster assignments}

\item{mu}{a T-by-K-by-d array of means}
}
\description{
Plot data colored by cluster assignment with cluster means
}