% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_pi}
\alias{plot_pi}
\title{Plot cluster populations (pi) over time}
\usage{
plot_pi(pi)
}
\arguments{
\item{pi}{A T-by-K array, with each row consisting of probabilities that sum to one}
}
\description{
Plot cluster populations (pi) over time
}