% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{kernel_em}
\alias{kernel_em}
\title{A Kernel-smoothed EM algorithm for a mixture of Gaussians that changes over time}
\usage{
kernel_em(
  y,
  K,
  hmu,
  hSigma,
  hpi,
  num_iter = 10,
  biomass = default_biomass(y),
  initial_fit = init_const(y, K, 50, 50)
)
}
\arguments{
\item{y}{length T list with \code{y[[t]]} being a n_t-by-d matrix}

\item{K}{number of components}

\item{hmu}{bandwidth for mu parameter}

\item{hSigma}{bandwidth for Sigma parameter}

\item{hpi}{bandwidth for pi parameter}

\item{num_iter}{number of iterations of EM to perform}

\item{biomass}{list of length T, where each element \code{biomass[[t]]} is a
numeric vector of length n_t containing the biomass (or count) of particles
in each bin}

\item{initial_fit}{a list of starting values for the parameters, responsibilities, and estimated cluster assignments}
}
\description{
\link{Add here}
}